requirements.txt
Django==5.0.14
PyMySQL==1.1.1
numpy==2.2.6
pandas==2.2.3
psycopg2-binary==2.9.10
matplotlib==3.10.1
